Unify and factor out try/finally unwinding.
What's going on here is that all user-level exceptions need to unwind
the same stack and go through the same handling paths. It is neatest and
hardest to miss cases when we raise all such exceptions through a single
type.

Interestingly, this leads to a special case for try-expressions
unwinding when ejectors are being raised. In this specific case, we say
that there is *not* any sealed value being thrown, and so even with some
hypothetical sealed value, we cannot run the catcher.

There isn't any stack tracing, but there ought to be.